services/menu_service/requirements-test.txt
# Tests — menu_service
# Instalar con: pip install -r requirements-test.txt

pytest==8.3.5
pytest-django==4.9.0
pytest-mock==3.14.0
pytest-cov==6.1.0
factory-boy==3.3.3
Faker==33.3.1